% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCodebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Create a Data Codebook}
\usage{
create_codebook(data, filename, output_dir = "auto", overwrite = FALSE)
}
\arguments{
\item{data}{The dataframe to export and parse}

\item{filename}{The file name (with no extension) to export the codebook & data file as}

\item{output_dir}{The directory to put the file, Default: the "intermediate" directory inside of open_data
using puddingR structure}

\item{overwrite}{Whether to overwrite the file if it already exists, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates an Rmd file using the parameterized Rmd template "codebook"
}
\details{
DETAILS
}
\examples{
\dontrun{
create_codebook(mtcars, "cars")
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
