% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFileStructure.R
\name{createFileStructure}
\alias{createFileStructure}
\title{createFileStructure}
\usage{
createFileStructure(directory = "analysis")
}
\arguments{
\item{directory}{Parent directory for the project to be created in. If the typed directory does not exist, it will be created., Default: 'analysis'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates a new R project with the folder structure used in Pudding R projects
}
\details{
DETAILS
}
\examples{
\dontrun{
createFileStructure("analysis")
}
}
\seealso{
\code{\link[usethis]{create_package}},\code{\link[usethis]{use_directory}}
 \code{\link[rstudioapi]{isAvailable}}
 \code{\link[purrr]{map2}}
}
