% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Create New Pudding Project}
\usage{
createProject(name = "analysis", title = NULL, folder = getwd(),
  dirs = c("assets", "assets/data/open_data", "plots",
  "assets/data/processed_data", "assets/data/raw_data", "reports", "rmds"),
  packagedeps = "packrat", reset = TRUE, open = TRUE)
}
\arguments{
\item{name}{File name for the project. Default: "analysis"}

\item{title}{What the project does, one line, title case, Default: NULL}

\item{folder}{Folder under which to create the new project, Default: getwd()}

\item{dirs}{Character vector of new directories to create, Default: c("assets", "functions", "open_data", "plots", "processed_data",
"raw_data", "reports", "rmds", "rscripts")}

\item{packagedeps}{Set which tool you would like to use for package reproducibility, Default: 'packrat'}

\item{reset}{Whether to reset the active project to your new project, Default: TRUE}

\item{open}{Whether the new project should open automatically, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates a new R project and generates the folders needed for consistent Pudding analysis
}
\details{
DETAILS
}
\examples{
\dontrun{
folder <- tempdir()
createProject(
  name = "pudding-analysis", title = "My Next Great Story",
  folder = folder,
  packagedeps = "checkpoint",
  reset = TRUE,
  open = FALSE
)
list.files(file.path(folder, "pudding-analysis"))
unlink(file.path(folder, "pudding-analysis"))
}
}
\seealso{
\code{\link[usethis]{create_package}},\code{\link[usethis]{proj_utils}},\code{\link[usethis]{use_description}}
 \code{\link[rstudioapi]{isAvailable}},\code{\link[rstudioapi]{projects}}
 \code{\link[desc]{desc_set}}
}
