% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{export_data}
\alias{export_data}
\title{Export Processed Data}
\usage{
export_data(data, filename, location = "processed", directory = "auto",
  format = "csv", na = "")
}
\arguments{
\item{data}{data frame to export}

\item{filename}{the name of the exported file, without the extension}

\item{location}{If using the puddingR file directory, automatically place the file inside either the open_data folder ("open")
the processed_data ("processed"), or outside of the R directory, and in the larger Pudding starter template ("js"), Default: 'processed'}

\item{directory}{Directory of the exported file. If set to "auto", this assumes that the project follows the
puddingR structure. Otherwise, the "location" parameter is overwritten by the directory in this argument, Default: 'auto'}

\item{format}{file output format (either "csv", "json", or "tsv"), Default: 'csv'}

\item{na}{How to export NA values, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Export any dataframe that has been processed in R
}
\details{
DETAILS
}
\examples{
\dontrun{
# export a csv using the puddingR template
export_data(mtcars, "cars")

# export as json using the puddingR template
export_data(mtcars, "cars", format = "json")

# output csv to a location not using the puddingR template
export_data(mtcars, "cars", directory = "data/my_data/")
}
}
\seealso{
\code{\link[dplyr]{case_when}}
}
