% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitDataReadme.R
\name{knit_data_readme}
\alias{knit_data_readme}
\title{Knit Codebooks to Single README}
\usage{
knit_data_readme(input_dir = "auto", filename = "README",
  output_dir = "auto")
}
\arguments{
\item{input_dir}{File directory of the existing codebooks, Default: "auto", assumes puddingR file directory
and uses the "intermediate" folder inside "open_data"}

\item{filename}{Name of output file, Default: 'README'}

\item{output_dir}{File directory for final README, Default: "auto", assumes puddingR file directory, outputs to
the "open_data" directory}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Combine all other generated codebooks into a single README file for a data repository
}
\details{
DETAILS
}
\examples{
\dontrun{
# using puddingR file structure
knit_data_readme()

# using other file structure
knit_data_readme(directory = "my_codebooks", output_dir = "public_data")
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
