% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddingR.R
\name{puddingR}
\alias{puddingR}
\title{puddingR}
\usage{
puddingR(cachePath = "~\\\\")
}
\arguments{
\item{cachePath}{path of the cache}
}
\description{
puddingR
}
\examples{
\dontrun{
puddingR()
}
}
\keyword{puddingR}

