% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCodebook.R
\name{render_codebook}
\alias{render_codebook}
\title{Render Data Codebook}
\usage{
render_codebook(data, filename)
}
\arguments{
\item{data}{The dataframe to export and parse}

\item{filename}{The file name (with no extension) to export the codebook & data file as}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates an Rmd file using the parameterized Rmd template "codebook"
}
\details{
DETAILS
}
\seealso{
\code{\link[rmarkdown]{render}}
}
