% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFileStructure.R, R/functions.R
\name{createFileStructure}
\alias{createFileStructure}
\title{createFileStructure}
\usage{
createFileStructure()

createFileStructure()
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
Creates a new R project with the folder structure used in Pudding R projects

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
createFileStructure()
}
}
\seealso{
\code{\link[usethis]{create_package}},\code{\link[usethis]{use_directory}}
 \code{\link[rstudioapi]{isAvailable}}
 \code{\link[base]{c}}
 \code{\link[purrr]{map2}}

\code{\link[usethis]{create_package}},\code{\link[usethis]{use_directory}}
 \code{\link[rstudioapi]{isAvailable}}
 \code{\link[base]{c}}
 \code{\link[purrr]{map2}}
}
