% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCode.R
\name{export_code}
\alias{export_code}
\title{Export several code chunks from Rmd to R}
\usage{
export_code(file, toKeep, outputFile, outputDir = "auto")
}
\arguments{
\item{file}{The .Rmd file to parse}

\item{toKeep}{A character vector of the code chunk labels to keep}

\item{outputFile}{The filename to write to (no extension)}

\item{outputDir}{The directory to write the file to (ending with trailing slash), Default: "auto"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
For use in creating a data repo, indicate all the code chunks used to create a specific subset of data
}
\details{
DETAILS
}
\examples{
\dontrun{
# In the puddingR default directory
export_code("analysis.Rmd", toKeep = c("load_data", "find_average"), outputFile = "average")

#' # If not in the puddingR default directory
export_code("analysis.Rmd", toKeep = c("load_data", "find_average"),
outputFile = "average", outputDir = "myFolder/")
}
}
\seealso{
\code{\link[dplyr]{case_when}}
 \code{\link[knitr]{knit}},\code{\link[knitr]{read_chunk}}
 \code{\link[purrr]{map}}
}
