% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitDataReadme.R
\name{knit_data_readme}
\alias{knit_data_readme}
\title{Knit Codebooks to Single README}
\usage{
knit_data_readme(directory = paste0(getwd(),
  "/assets/data/open_data/intermediate/"), filename = "README",
  output_dir = paste0(getwd(), "/assets/data/open_data/"))
}
\arguments{
\item{directory}{File directory of the existing codebooks, Default: paste0(getwd(), "/assets/data/open_data/intermediate/")}

\item{filename}{Name of output file, Default: 'README'}

\item{output_dir}{File directory for final README, Default: paste0(getwd(), "/assets/data/open_data/")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Combine all other generated codebooks into a single README file for a data repository
}
\details{
DETAILS
}
\seealso{
\code{\link[rmarkdown]{render}}
}
